% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_query_url.R
\name{construct_query_url}
\alias{construct_query_url}
\title{Construct query URL for Ember data API}
\usage{
construct_query_url(
  endpoint = c("electricity-generation", "power-sector-emissions", "electricity-demand",
    "carbon-intensity"),
  temporal_resolution = c("yearly", "monthly"),
  year = NA,
  min_date = "2015",
  max_date = "2023",
  entity = "all",
  api_key = Sys.getenv("EMBER_API_KEY")
)
}
\arguments{
\item{endpoint}{Default = "electricity-generation"}

\item{temporal_resolution}{"yearly" (default) or "monthly"}

\item{year}{(optional) If specified, retrieve data for a single year}

\item{min_date}{Mininum date to retrieve data for. YYYY or YYYY-MM format. Default=2015}

\item{max_date}{Maximum date to retrieve data for. YYYY or YYYY-MM format. Default=2023}

\item{entity}{List of comma-separated country(s) or region(s) to return data for. Default is all (no filter)}

\item{api_key}{Default is Sys.getenv("EMBER_API_KEY")}
}
\value{
query_url API query URL for specified parameters
}
\description{
Construct query URL for Ember data API
}
