% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emberr-package.R
\docType{package}
\name{emberr-package}
\alias{emberr}
\alias{emberr-package}
\title{emberr: R API Wrapper For Ember Data}
\description{
Provides R functions to retrieve data from the Ember Climate API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andypicke/emberr}
  \item Report bugs at \url{https://github.com/andypicke/emberr/issues}
}

}
\author{
\strong{Maintainer}: Andy Pickering \email{andypicke@gmail.com} (\href{https://orcid.org/0000-0002-5724-7042}{ORCID})

}
\keyword{internal}
