% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_electricity_generation.R
\name{get_electricity_generation}
\alias{get_electricity_generation}
\title{Get Ember electricity-generation data from API}
\usage{
get_electricity_generation(
  temporal_resolution = "yearly",
  min_date = "2015",
  max_date = "2023",
  entity = "all",
  api_key = Sys.getenv("EMBER_API_KEY")
)
}
\arguments{
\item{temporal_resolution}{"yearly" (default) or "monthly"}

\item{min_date}{Mininum date to retrieve data for. YYYY or YYYY-MM format. Default=2015}

\item{max_date}{Maximum date to retrieve data for. YYYY or YYYY-MM format. Default=2023}

\item{entity}{List of comma-separated country(s) or region(s) to return data for. Default is all (no filter)}

\item{api_key}{Default is Sys.getenv("EMBER_API_KEY")}
}
\value{
df Dataframe of requested data
}
\description{
Get Ember electricity-generation data from API
}
