% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ember_options.R
\name{get_ember_options}
\alias{get_ember_options}
\title{Get parameter options for Ember API}
\usage{
get_ember_options(
  endpoint = "electricity-generation",
  temporal_resolution = "yearly",
  param = "entity",
  api_key = Sys.getenv("EMBER_API_KEY")
)
}
\arguments{
\item{endpoint}{Default="electricity-generation"}

\item{temporal_resolution}{default ="yearly"}

\item{param}{Parameter to get options for}

\item{api_key}{Default is Sys.getenv("EMBER_API_KEY")}
}
\value{
options List of options for specified parameter
}
\description{
Get parameter options for Ember API
}
