% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_electricity_generation.R
\name{get_electricity_generation}
\alias{get_electricity_generation}
\title{Get Ember electricity-generation data from API}
\usage{
get_electricity_generation(
  temporal_resolution = "yearly",
  min_date = "2015",
  max_date = "2023",
  api_key = Sys.getenv("EMBER_API_KEY")
)
}
\arguments{
\item{temporal_resolution}{"yearly" (default) or "monthly"}

\item{min_date}{default=2015}

\item{max_date}{default=2023}

\item{api_key}{Default is Sys.getenv("EMBER_API_KEY")}
}
\value{
df Dataframe of requested data
}
\description{
Get Ember electricity-generation data from API
}
