% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ember_options.R
\name{get_ember_options}
\alias{get_ember_options}
\title{Get API parameter options for Ember dataset}
\usage{
get_ember_options(
  dataset = "electricity-generation",
  temporal_resolution = "yearly",
  filter_name = "entity",
  api_key = Sys.getenv("EMBER_API_KEY")
)
}
\arguments{
\item{dataset}{Default="electricity-generation"}

\item{temporal_resolution}{default ="yearly"}

\item{filter_name}{Parameter to get options for (ex: entity, series, date. Default = "entity"}

\item{api_key}{Default is Sys.getenv("EMBER_API_KEY")}
}
\value{
options List of options for specified parameter
}
\description{
Get API parameter options for Ember dataset
}
